filemanager finished

+++ actions/admin/filemanager.php
@@ -11,12 +11,12 @@ function returnBack() {
 if(isset($_POST["selected_files"]) && isset($_POST["delete"])) {
     if(count($_POST["selected_files"]) > 0) {
         foreach($_POST["selected_files"] as $file) {
-            unlink($file);
+            if(is_file($file)) unlink($file); else rmdir($file);
         }
         include_once("actions/admin/logger.php");
         returnBack();
     }
-} else header("location: /admin/fileManager?error=nothing_selected");
+}
 
 if(isset($_POST["addFile"]) && isset($_POST["fileName"])) {
     if(isset($_POST["backlink"])) {
@@ -26,9 +26,22 @@ if(isset($_POST["addFile"]) && isset($_POST["fileName"])) {
     include_once("actions/admin/logger.php");
     returnBack();
 }
+if(isset($_POST["addFolder"]) && isset($_POST["folderName"])) {
+    if(isset($_POST["backlink"])) {
+        $fullPath = $_SERVER['DOCUMENT_ROOT']."\\".$_POST["backlink"]."\\".$_POST["folderName"];
+    } else $fullPath = $_SERVER['DOCUMENT_ROOT']."\\".$_POST["folderName"];
+    mkdir($fullPath);
+    include_once("actions/admin/logger.php");
+    returnBack();
+}
 
 if(isset($_POST["saveFile"]) && isset($_POST["contents"]) && isset($_POST["filepath"])) {
     file_put_contents($_POST["filepath"], $_POST["contents"]);
+    include_once("actions/admin/logger.php");
+    returnBack();
+}
+if(isset($_POST["saveFolder"]) && isset($_POST["folderName"]) && isset($_POST["originalFolderName"])) {
+    rename($_POST["originalFolderName"], dirname($_POST["originalFolderName"])."\\".$_POST["folderName"]);
     include_once("actions/admin/logger.php");
     returnBack();
 }
